pyproject.toml
[project]
name = "mqtt-paho-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "paho-mqtt>=2.1.0",
    "protobuf>=3.19.1",
    "tahu>=0.1.2",
]
